requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
celery==5.5.3
redis==6.4.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.43
alembic==1.15.2
llama-index-core==0.12.42
pypdf==5.1.0
python-docx==1.1.2
